Read what a structured buffer holds
Recorded last round as not fixable: dcl_resource_structured carries a
stride and no type, so a StructuredBuffer was written as holding floats
whatever it held. A buffer of uints hashed with shifts and xors then reads
as float, which computes the wrong thing and does not compile.

The reflection data does say. fxc emits one constant buffer per structured
buffer, named after it, of type resource bind info, holding a single
variable called $Element whose type is the element type. The reader was
already walking those buffers and discarding both the name and the type -
the name was read from its offset and never kept, and D3DCbufferType had
only the two D3D10 values, so the third and fourth went unrecognised.

Two lines to keep the name, four to route it, and the resource carries its
element type. Where the reflection is silent the old guess stands.

It corrected integer_multiply too, whose buffer is uint and had been
written float, and that one stops regressing, so the ratchet loses an
entry as well.

The reason written down for not fixing this was accurate about the
instruction stream and wrong about the shader: the type was two chunks
away.

+++ HlslDecompiler.Tests/ShaderSources/cs_4_1/compute_hash.fx
@@ -1,5 +1,5 @@
-StructuredBuffer<float> src : register(t0);
-RWStructuredBuffer<float> dst : register(u0);
+StructuredBuffer<uint> src : register(t0);
+RWStructuredBuffer<uint> dst : register(u0);
 
 struct CS_IN
 {
